Restore all changes from weird repo
