+++ src/components/ChatBox.jsx
@@ -1,62 +1,19 @@
-import React, { useState, useContext, useEffect } from 'react';
-import { SocketContext } from '../context/SocketProvider';
+import React from 'react';
+import useChat from '../hooks/useChat';
 import style from './ChatBox.css';
 
 const ChatBox = () => {
-  const socket = useContext(SocketContext);
-
-  const [messages, setMessages] = useState([]);
-  // tracking whether there are messages from other users that the client player has not read
-  const [unreadMessages, setUnreadMessages] = useState(0);
-
-  const [inputMessage, setInputMessage] = useState('');
-  const [collapsed, setCollapsed] = useState(true);
-
-  const handleFormSubmit = e => {
-    e.preventDefault();
-
-    if(inputMessage) {
-      
-      const messageObj = {
-        sender: socket.id,
-        text: inputMessage
-      };
-
-      socket.emit('client message', messageObj);
-      setMessages(prev => [...prev, messageObj]);
-
-      setInputMessage('');
-    }
-  };
-
-  const handleIncomingMessage = message => {
-    const audio = document.querySelector('#chat-audio');
-    audio.volume = 0.1;
-    audio.play();
-
-    setMessages(prev => [...prev, message]);
-    if(collapsed) setUnreadMessages(unreadMessages => unreadMessages + 1);
-  };
-
-  const handleExpandClick = () => {
-    setCollapsed(false);
-    setUnreadMessages(0);
-  };
-
-  useEffect(() => {
-    if(messages.length && messages[messages.length - 1].sender === socket.id) {
-      const messageList = document.getElementById('messageList');
-      messageList.scrollTo(0, messageList.scrollHeight, { behavior: 'smooth' });
-    }
-  }, [messages]);
-
-  useEffect(() => {
-    socket.on('socket message', handleIncomingMessage);
-    return () => {
-      // socket.off('current users', handleCurrentUsers);
-      socket.off('socket message', handleIncomingMessage);
-    };
-  }, [socket]);
+  const {
+    clientId,
+    messages,
+    unreadMessages,
+    inputMessage,
+    setInputMessage,
+    collapsed,
+    setCollapsed,
+    handleExpandClick,
+    handleFormSubmit
+  } = useChat();
 
   return (
     <div>
@@ -88,7 +45,7 @@ const ChatBox = () => {
           
           <ul aria-label="message list" className={style.messageList} id="messageList">
             {messages.map((message, index) => {
-              return message.sender === socket.id
+              return message.sender === clientId
                 ? <li
                   key={message + '-' + index}
                   className={style.selfMessage}>
@@ -115,12 +72,9 @@ const ChatBox = () => {
             />
             <button aria-label="message submit">Send</button>
           </form>
-  
         </div>
       }
     </div>
-    
-    
   );
 };
