+++ src/components/profile/CompanyProfile.css
@@ -38,6 +38,31 @@
   color: #3f8a72;
 }
 
+.cardInner {
+  transition: transform .8s;
+  transform-style: preserve-3d;
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.cardFront, .cardBack {
+  cursor: pointer;
+  position: absolute;
+  backface-visibility: hidden;
+  -webkit-backface-visibility: hidden;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: 2%;
+}
+
+.cardBack, .cardInner.flip {
+  transform: rotateY(180deg);
+}
+
 .news {
   display: flex;
   flex-direction: column;
@@ -73,10 +98,6 @@
 
 .objectPulse {
   animation: object-pulse-animation 5s infinite;
-}
-
-.objectPulse:hover {
-  cursor: pointer;
 }
 .glowObjectChallengeEffect {
   transform: rotate(180deg);
