+++ src/components/profile/CompanyProfile.jsx
@@ -7,6 +7,7 @@ import { GameStateContext } from '../../context/GameStateProvider';
 const CompanyProfile = ({ glowingObjectState, glowChangeHandler }) => {
   const { showGhostStory, handleStoryClick } = useGhostStory();
   const { points, failedTimeouts } = useContext(GameStateContext);
+
   return (
     <section className={style.companyProfile}>
       <article className={style.story}>
@@ -16,29 +17,33 @@ const CompanyProfile = ({ glowingObjectState, glowChangeHandler }) => {
             src={points >= 10 ? '/assets/thumbs3.jpg' : (failedTimeouts >= 3 ? '/assets/thumbs2.jpg' : '/assets/thumbs1.jpg')} alt="gim image" />
         </figure>
 
-        {showGhostStory ? (
-          <article onClick={handleStoryClick}>
-            <h2>A Different Story</h2>
-            <p>
-              I don&apos;t know what happened. An accident, maybe? <br />
-              <br />
-              But this isn&apos;t what I thought it would feel like. <br />
-              <br />
-              I was at work. Late. <br />
-            </p>
-          </article>
-        ) : (
-          <article onClick={handleStoryClick}>
+        
+        <div className={showGhostStory ? `${style.cardInner} ${style.flip}` : style.cardInner}>
+          <article 
+            className={style.cardFront} 
+            onClick={handleStoryClick}>
             <h2>The GIM Story</h2>
             <p>
-              General Industrial Manufacturing Company is a three-generation,
+              General Industrial Manufacturing is a three-generation,
               family-owned and operated, precision machining supplier. GIM has
               earned a reputation as a supplier of high quality, low cost
               precision parts with on-time deliveries. We are a top quality team
               of employees who work well together.
             </p>
           </article>
-        )}
+
+          <article 
+            className={style.cardBack}
+            onClick={handleStoryClick}>
+            <h2>A Different Story</h2>
+            <p>
+              Robin Smith was a longtime employee in the GIM IT department. While performing maintenance in the server room, something unusual happened. <br /> <br />
+              <code>I don&apos;t know what happened, an accident maybe? </code>
+              <code>... can you help?</code>
+            </p>
+          </article>
+        </div>
+
 
       </article>
